Improve new Layout class
* Now in namespace UI::Wizards::Layout
* LayoutConfig was dropped
* Added Wizard#OpenWithLayout
