+++ library/control/src/modules/ProductFeatures.rb
@@ -189,7 +189,7 @@ module Yast
 
     # Initialize the features structure if needed
     #
-    # Note that the values are normally read from a control file or set to default values. But, in normal
+    # @note The values are normally read from a control file or set to default values. But, in normal
     # mode, the values are taken from /etc/YaST2/ProductFeatures. Also note that for firstboot mode, the
     # values are taken from the firstboot.xml control file.
     #
