+++ library/wizard/test/wizard_test.rb
@@ -20,6 +20,7 @@
 # find current contact information at www.suse.com.
 
 require_relative "test_helper"
+require "ui/wizards/layout"
 
 Yast.import "Wizard"
 
@@ -76,6 +77,74 @@ describe Yast::Wizard do
       allow(Yast::UI).to receive(:UserInput).and_return(:cancel)
 
       expect(subject.UserInput).to eq :back
+    end
+  end
+
+  describe ".OpenWithLayout" do
+    let(:layout) { ::UI::Wizards::Layout.with_steps }
+
+    context "when the banner is enabled in the layout" do
+      before do
+        layout.show_banner
+      end
+
+      it "shows the banner" do
+        expect(Yast::UI).to receive(:SetProductLogo).with(true)
+
+        subject.OpenWithLayout(layout)
+      end
+    end
+
+    context "when the banner is disabled in the layout" do
+      before do
+        layout.hide_banner
+      end
+
+      it "hides the banner" do
+        expect(Yast::UI).to receive(:SetProductLogo).with(false)
+
+        subject.OpenWithLayout(layout)
+      end
+    end
+
+    context "when the layout mode is steps" do
+      let(:layout) { ::UI::Wizards::Layout.with_steps }
+
+      it "opens a wizard with steps" do
+        expect(Yast::Wizard).to receive(:OpenNextBackStepsDialog)
+
+        subject.OpenWithLayout(layout)
+      end
+    end
+
+    context "when the layout mode is tree" do
+      let(:layout) { ::UI::Wizards::Layout.with_tree }
+
+      it "opens a wizard with tree" do
+        expect(Yast::Wizard).to receive(:OpenTreeNextBackDialog)
+
+        subject.OpenWithLayout(layout)
+      end
+    end
+
+    context "when the layout mode is title-on-left" do
+      let(:layout) { ::UI::Wizards::Layout.with_title_on_left }
+
+      it "opens a wizard with title on left" do
+        expect(Yast::Wizard).to receive(:OpenLeftTitleNextBackDialog)
+
+        subject.OpenWithLayout(layout)
+      end
+    end
+
+    context "when the layout mode is title-on-top" do
+      let(:layout) { ::UI::Wizards::Layout.with_title_on_top }
+
+      it "opens a wizard with title on top" do
+        expect(Yast::Wizard).to receive(:OpenNextBackDialog)
+
+        subject.OpenWithLayout(layout)
+      end
     end
   end
 end
